Improve deletion of previous user profile image
This commit refactors the `upload_image` function in `views.py` to improve the deletion of the previous user profile image. Previously, the code would always attempt to delete the previous image, even if it was the default profile picture. This resulted in unnecessary API calls to the S3 bucket.

Now, the code first checks if the user's profile picture is the default image. If it is, the deletion process is skipped. However, if the profile picture is a custom image, the previous image URL is retrieved and the corresponding image file is deleted from the S3 bucket.

This change improves the efficiency of the image deletion process and reduces unnecessary API calls.

+++ app_project/airfinn/views.py
@@ -1180,15 +1180,17 @@ def upload_image(request):
         except User.DoesNotExist:
             return JsonResponse({'error': 'User not found'}, status=404)
         
-        # Delete the previous image
-        previous_image_url = user.profile_picture_url
-        
-        # Extract the filename from the URL
-        if previous_image_url:
-            previous_image_filename = previous_image_url.split('/')[-1]
+        # Delete the previous image if the picture is not the default image
+        if user.profile_picture_url != 'https://rentopia-images.dybedahlserver.net/default_profile_picture.jpg':
+            previous_image_url = user.profile_picture_url
             
-            # Delete the previous image from the S3 bucket
-            delete_image(previous_image_filename)
+        else:
+            # Extract the filename from the URL
+            if previous_image_url:
+                previous_image_filename = previous_image_url.split('/')[-1]
+                
+                # Delete the previous image from the S3 bucket
+                delete_image(previous_image_filename)
         
         # Initialize the S3 client with your credentials and endpoint
         ACCOUNT_ID = os.getenv('ACCOUNT_ID')
